requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
aiohttp==3.8.4
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
fake-useragent==1.4.0